fix register allocation
This is a very interesting bug, we compare vertex by reg only. So in the
union operation, if I have two vertices

1. `{ value = x0; adjacency = [] }`
2. and `{ value = x0; adjacency = [ x1 ] }`

the node would be dropped(horrifying).

This fix changes the order of union, and hence, the newer vertex is in.

But this still has a problem that if I got

1. `{ value = x0; adjacency = [ x1 ] }`
2. `{ value = x0; adjacency = [ x2 ] }`

The correct result for this is having a new vertex: `{ value = x0; adjacency = [ x1; x2 ] }`

Signed-off-by: Lîm Tsú-thuàn <dannypsnl@gmail.com>